Time series changed from long to int on storage (only seconds) Disabled remote actors for testing purposes

+++ src/main/scala/au/csiro/ict/InputProcessingActor.scala
@@ -20,7 +20,7 @@ class InputProcessingWorker extends Actor{
       println("I am a worker and now I am processing:"+queueName)
       val Array(_,nid,sid)=queueName.split("@")
       while(queue.llen().isDefined){
-        store.addNodeData(nid,Map(sid->parse[Map[Long,String]](queue.lrange(queueName,0,0).head.head.get)))
+        store.addNodeData(nid,Map(sid->parse[Map[Int,String]](queue.lrange(queueName,0,0).head.head.get)))
         Cache.queue.lpop()
       }
       sender ! Done(queueName)
